rename all mentions of ingest* to to index*
