+++ examples/Cargo.toml
@@ -16,24 +16,24 @@ serde_json = "1.0"
 spider = "1.98"
 
 [[example]]
-name = "ingest-codebase"
-path = "ingest_codebase.rs"
+name = "index-codebase"
+path = "index_codebase.rs"
 
 [[example]]
 name = "fastembed"
 path = "fastembed.rs"
 
 [[example]]
-name = "ingest-redis"
-path = "ingest_into_redis.rs"
+name = "index-redis"
+path = "index_into_redis.rs"
 
 [[example]]
-name = "ingest-markdown-metadata"
-path = "ingest_markdown_lots_of_metadata.rs"
+name = "index-markdown-metadata"
+path = "index_markdown_lots_of_metadata.rs"
 
 [[example]]
-name = "scraping-ingest"
-path = "scraping_ingest_to_markdown.rs"
+name = "scraping-index"
+path = "scraping_index_to_markdown.rs"
 
 [[example]]
 name = "aws-bedrock"
