+++ examples/aws_bedrock.rs
@@ -13,7 +13,7 @@
 //! [AWS Bedrock documentation]: https://docs.aws.amazon.com/bedrock/
 
 use swiftide::{
-    ingestion, integrations, loaders::FileLoader, persist::MemoryStorage, transformers,
+    indexing, integrations, loaders::FileLoader, persist::MemoryStorage, transformers,
 };
 
 #[tokio::main]
@@ -27,7 +27,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let memory_storage = MemoryStorage::default();
 
-    ingestion::Pipeline::from_loader(FileLoader::new("./README.md"))
+    indexing::Pipeline::from_loader(FileLoader::new("./README.md"))
         .log_nodes()
         .then_chunk(transformers::ChunkMarkdown::from_chunk_range(100..512))
         .then(transformers::MetadataSummary::new(aws_bedrock.clone()))
