+++ swiftide/src/indexing/indexing_stream.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::from_over_into)]
 #![cfg(not(tarpaulin_include))]
-//! This module defines the `IngestionStream` type, which is used for handling asynchronous streams of `IngestionNode` items in the ingestion pipeline.
+//! This module defines the `IngestionStream` type, which is used for handling asynchronous streams of `IngestionNode` items in the indexing pipeline.
 
 use anyhow::Result;
 use futures_util::stream::{self, Stream};
