+++ swiftide/src/lib.rs
@@ -1,6 +1,6 @@
 //! Swiftide - Document and code indexation for retrieval augmented generation
 //!
-//! Swiftide is a straightforward, easy-to-use, easy-to-extend asynchronous file ingestion and processing system. It is designed to be used in a RAG (Retrieval Augmented Generation) system. It is built to be fast and efficient, with a focus on parallel processing and asynchronous operations.
+//! Swiftide is a straightforward, easy-to-use, easy-to-extend asynchronous file indexing and processing system. It is designed to be used in a RAG (Retrieval Augmented Generation) system. It is built to be fast and efficient, with a focus on parallel processing and asynchronous operations.
 //!
 //! Part of the bosun.ai project. An upcoming platform for autonomous code improvement.
 //!
@@ -27,9 +27,9 @@ pub use type_aliases::*;
 /// Deprecated re-export of `indexing`, use that instead.
 #[deprecated(
     since = "0.6.0",
-    note = "Renamed references of Ingestion to Indexing for more appropriate naming. Will be removed in a future release."
+    note = "Renamed references of Ingestion to Ingestion for more appropriate naming. Will be removed in a future release."
 )]
-pub mod ingestion {
+pub mod ingestion  {
     pub use crate::indexing::*;
 
     pub use crate::indexing::IndexingStream as IngestionStream;
