+++ swiftide/src/transformers/chunk_markdown.rs
@@ -1,5 +1,5 @@
 //! Chunk markdown content into smaller pieces
-use crate::{ingestion::IndexingStream, ingestion::Node, ChunkerTransformer};
+use crate::{indexing::IndexingStream, indexing::Node, ChunkerTransformer};
 use async_trait::async_trait;
 use derive_builder::Builder;
 use text_splitter::{Characters, MarkdownSplitter};
