+++ swiftide/src/transformers/metadata_keywords.rs
@@ -1,7 +1,7 @@
 //! Extract keywords from a node and add them as metadata
 use std::sync::Arc;
 
-use crate::{ingestion::Node, SimplePrompt, Transformer};
+use crate::{indexing::Node, SimplePrompt, Transformer};
 use anyhow::Result;
 use async_trait::async_trait;
 use derive_builder::Builder;
